基于SDL实现RichEdit的功能（调整数据结构，减少内存占有率，未完待续）

+++ duilib/Control/RichEdit_SDL.h
@@ -825,7 +825,7 @@ private:
 
     /** 获取当前绘制文字的属性
     */
-    uint32_t GetTextStyle() const;
+    uint16_t GetTextStyle() const;
 
     /** 将文本生成可绘制的格式
     */
